+++ lib/grape/roar/extensions/relations/validations.rb
@@ -1,0 +1,3 @@
+require 'grape/roar/extensions/relations/validations/misc'
+require 'grape/roar/extensions/relations/validations/active_record'
+require 'grape/roar/extensions/relations/validations/mongoid'