+++ spec/extensions/relations/validations/active_record_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 require 'active_record'
 
-describe Grape::Roar::Extensions::RelationalModels::Validations::ActiveRecord do
+describe Grape::Roar::Extensions::Relations::Validations::ActiveRecord do
   let(:model_klass) { double }
 
   subject do
@@ -39,7 +39,7 @@ describe Grape::Roar::Extensions::RelationalModels::Validations::ActiveRecord do
       it 'properly validates the relation' do
         expect(subject.send("#{relation}_valid?", :test)).to eql(true)
         expect { subject.send("#{relation}_valid?", :fail) }.to raise_error(
-          Grape::Roar::Extensions::RelationalModels::\
+          Grape::Roar::Extensions::Relations::\
           Exceptions::InvalidRelationError
         )
       end